Fixed package names in moved files.
