+++ test/integration/artifactcache_int_test.go
@@ -1,4 +1,4 @@
-package artifactcache
+package integration
 
 import (
 	"os"
